Isolating the logic change for the replacement image optional to the refactor
